refactor(frontend/context): use the context directory to clean up main.tsx
