+++ app/frontend/src/hooks/auth.ts
@@ -1,5 +1,5 @@
-import { useContext } from "react";
 import { AuthContext } from "@/context/auth-provider"; // Adjust the import path as necessary
+import { useContext } from "react";
 
 export function useAuth() {
 	const user = useContext(AuthContext);
